[feature/image_experiment] 코드 정리
