+++ app/src/main/java/com/teamfillin/fillin/data/service/HomeService.kt
@@ -3,10 +3,8 @@ package com.teamfillin.fillin.data.service
 import com.teamfillin.fillin.data.response.BaseResponse
 import com.teamfillin.fillin.data.response.ResponseNewPhotoInfo
 import com.teamfillin.fillin.data.response.ResponseUserInfo
-import okhttp3.MultipartBody
-import okhttp3.RequestBody
 import retrofit2.Call
-import retrofit2.http.*
+import retrofit2.http.GET
 
 interface HomeService {
     @GET("photo/latest")
@@ -14,11 +12,4 @@ interface HomeService {
 
     @GET("user")
     fun getUser(): Call<BaseResponse<ResponseUserInfo>>
-
-    @Multipart
-    @POST("photo")
-    fun postImage(
-        @PartMap body: HashMap<String, RequestBody>,
-        @Part file: MultipartBody.Part?
-    ): Call<BaseResponse<Unit>>
 }
