+++ app/src/main/java/com/teamfillin/fillin/presentation/filmroll/add/AddCompleteDialog.kt
@@ -1,4 +1,4 @@
-package com.teamfillin.fillin.presentation
+package com.teamfillin.fillin.presentation.filmroll.add
 
 import android.app.Dialog
 import android.content.Context
