+++ app/src/main/java/com/teamfillin/fillin/presentation/home/HomeActivity.kt
@@ -3,7 +3,6 @@ package com.teamfillin.fillin.presentation.home
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import android.util.Log
 import androidx.core.view.isVisible
 import androidx.lifecycle.lifecycleScope
 import com.naver.maps.geometry.LatLng
@@ -18,21 +17,16 @@ import com.teamfillin.fillin.core.base.BindingActivity
 import com.teamfillin.fillin.core.content.receive
 import com.teamfillin.fillin.core.context.toast
 import com.teamfillin.fillin.core.view.setOnSingleClickListener
-import com.teamfillin.fillin.data.response.BaseResponse
 import com.teamfillin.fillin.data.response.ResponseNewPhotoInfo
-import com.teamfillin.fillin.data.response.ResponseUserInfo
 import com.teamfillin.fillin.data.service.HomeService
 import com.teamfillin.fillin.data.service.StudioService
 import com.teamfillin.fillin.databinding.ActivityHomeBinding
-import com.teamfillin.fillin.presentation.AddPhotoActivity
+import com.teamfillin.fillin.presentation.filmroll.add.AddPhotoActivity
 import com.teamfillin.fillin.presentation.MyPageActivity
 import com.teamfillin.fillin.presentation.filmroll.FilmRollActivity
 import com.teamfillin.fillin.presentation.map.StudioMapActivity
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
 import retrofit2.await
 import timber.log.Timber
 import javax.inject.Inject
